+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Use shared SSH keypair for remote builds instead of host keys
 - Use `vars` to store `syncthing` GUI password
 - Use `vars` to store Tailscale auth key
 - Added `gaia`